Refactor the handling of the ueberauth auth struct in the controller
