+++ web/router.ex
@@ -23,7 +23,7 @@ defmodule UeberauthExample.Router do
 
     get "/:provider", AuthController, :request
     get "/:provider/callback", AuthController, :callback
-    post "/identity/callback", AuthController, :identity_callback
+    post "/:provider/callback", AuthController, :callback
     delete "/logout", AuthController, :delete
   end
